+++ src/internet/helper/internet-stack-helper.cc
@@ -290,6 +290,12 @@ InternetStackHelper::InstallAll() const
 void
 InternetStackHelper::CreateAndAggregateObjectFromTypeId(Ptr<Node> node, const std::string typeId)
 {
+    TypeId tid = TypeId::LookupByName(typeId);
+    if (node->GetObject<Object>(tid))
+    {
+        return;
+    }
+
     ObjectFactory factory;
     factory.SetTypeId(typeId);
     Ptr<Object> protocol = factory.Create<Object>();
@@ -301,13 +307,7 @@ InternetStackHelper::Install(Ptr<Node> node) const
 {
     if (m_ipv4Enabled)
     {
-        if (node->GetObject<Ipv4>())
-        {
-            NS_FATAL_ERROR("InternetStackHelper::Install (): Aggregating "
-                           "an InternetStack to a node with an existing Ipv4 object");
-            return;
-        }
-
+        /* IPv4 stack */
         CreateAndAggregateObjectFromTypeId(node, "ns3::ArpL3Protocol");
         CreateAndAggregateObjectFromTypeId(node, "ns3::Ipv4L3Protocol");
         CreateAndAggregateObjectFromTypeId(node, "ns3::Icmpv4L4Protocol");
@@ -318,22 +318,19 @@ InternetStackHelper::Install(Ptr<Node> node) const
             arp->SetAttribute("RequestJitter",
                               StringValue("ns3::ConstantRandomVariable[Constant=0.0]"));
         }
+
         // Set routing
         Ptr<Ipv4> ipv4 = node->GetObject<Ipv4>();
-        Ptr<Ipv4RoutingProtocol> ipv4Routing = m_routing->Create(node);
-        ipv4->SetRoutingProtocol(ipv4Routing);
+        if (!ipv4->GetRoutingProtocol())
+        {
+            Ptr<Ipv4RoutingProtocol> ipv4Routing = m_routing->Create(node);
+            ipv4->SetRoutingProtocol(ipv4Routing);
+        }
     }
 
     if (m_ipv6Enabled)
     {
         /* IPv6 stack */
-        if (node->GetObject<Ipv6>())
-        {
-            NS_FATAL_ERROR("InternetStackHelper::Install (): Aggregating "
-                           "an InternetStack to a node with an existing Ipv6 object");
-            return;
-        }
-
         CreateAndAggregateObjectFromTypeId(node, "ns3::Ipv6L3Protocol");
         CreateAndAggregateObjectFromTypeId(node, "ns3::Icmpv6L4Protocol");
         if (m_ipv6NsRsJitterEnabled == false)
@@ -345,9 +342,11 @@ InternetStackHelper::Install(Ptr<Node> node) const
         }
         // Set routing
         Ptr<Ipv6> ipv6 = node->GetObject<Ipv6>();
-        Ptr<Ipv6RoutingProtocol> ipv6Routing = m_routingv6->Create(node);
-        ipv6->SetRoutingProtocol(ipv6Routing);
-
+        if (!ipv6->GetRoutingProtocol())
+        {
+            Ptr<Ipv6RoutingProtocol> ipv6Routing = m_routingv6->Create(node);
+            ipv6->SetRoutingProtocol(ipv6Routing);
+        }
         /* register IPv6 extensions and options */
         ipv6->RegisterExtensions();
         ipv6->RegisterOptions();
@@ -357,9 +356,17 @@ InternetStackHelper::Install(Ptr<Node> node) const
     {
         CreateAndAggregateObjectFromTypeId(node, "ns3::TrafficControlLayer");
         CreateAndAggregateObjectFromTypeId(node, "ns3::UdpL4Protocol");
-        node->AggregateObject(m_tcpFactory.Create<Object>());
-        Ptr<PacketSocketFactory> factory = CreateObject<PacketSocketFactory>();
-        node->AggregateObject(factory);
+        // note: the following could be changed to
+        // CreateAndAggregateObjectFromTypeId(node, "ns3::TcpL4Protocol");
+        if (!node->GetObject<TcpL4Protocol>())
+        {
+            node->AggregateObject(m_tcpFactory.Create<Object>());
+        }
+        if (!node->GetObject<PacketSocketFactory>())
+        {
+            Ptr<PacketSocketFactory> factory = CreateObject<PacketSocketFactory>();
+            node->AggregateObject(factory);
+        }
     }
 
     if (m_ipv4Enabled)